+++ R/tar_seed_create.R
@@ -28,26 +28,14 @@
 #'   L'Ecuyer et al. (2017) \doi{10.1016/j.matcom.2016.05.005}
 #'   under "A single RNG with a 'random' seed for each stream" (Section 4:
 #'   under "How to produce parallel streams and substreams").
-#'
 #'   `targets` and `tarchetypes` take the approach discussed in the
-#'   aforementioned section of the paper, where [tar_seed_create()] plays the
-#'   role of the upstream pseudo-random number generator (RNG) that produces
-#'   seeds for the subsequent parallel streams. Specifically,
-#'   [tar_seed_create()] acts as a counter-based RNG,
-#'   where the output function is the SHA512 hash algorithm.
-#'   (See "counter-based RNGs" under Section 3: "Main classes of RNGs
-#'   for simulation".) There are two notable quirks, but they are innocuous:
-#'
-#'     1. Instead of a recursive integer counter, [tar_seed_create()] accepts
-#'        the name of the target as the argument.
-#'        Because target names map to small integers, this is equivalent to
-#'        traversing the RNG sequence in variable small jumps. This does not
-#'        affect the period length, and the argument in Section 4 of
-#'        L'Ecuyer et al. (2017) still holds.
-#'     2. [tar_seed_create()] converts the SHA512 hash into an
-#'        integer using `digest::digest2int()`. This is necessary
-#'        because integers in R can only be 32 bits. `set.seed()`
-#'        cannot accept anything greater.
+#'   aforementioned section of the paper using the
+#'   `secretbase` package by Charlie Gao (2024) \doi{10.5281/zenodo.10553140}.
+#'   To generate the 32-bit integer `seed` argument of `set.seed()`
+#'   for each target, `secretbase` generates a cryptographic SHA3 hash
+#'   and robustly converts it to 32-bit output using the SHAKE256
+#'   extendable output function (XOF). `secretbase` uses algorithms from
+#'   the `Mbed TLS` C library.
 #' @return Integer of length 1, the target seed.
 #' @param name Character of length 1, target name.
 #' @param global_seed Integer of length 1, the overarching global
@@ -56,10 +44,13 @@
 #'   Set to `NA` to disable seed setting in `targets` and make
 #'   [tar_seed_create()] return `NA_integer_`.
 #' @references
-#'   Pierre L'Ecuyer, David Munger, Boris Oreshkin, and Richard Simard
-#'   (2017). Random numbers for parallel computers: Requirements and methods,
-#'   with emphasis on GPUs. Mathematics and Computers in Simulation,
-#'   135, 3-17. \doi{10.1016/j.matcom.2016.05.005}.
+#'   * Gao C (2024). `secretbase`: Cryptographic Hash and
+#'     Extendable-Output Functions. R package version 0.1.0,
+#'     \doi{10.5281/zenodo.10553140}.
+#'   * Pierre L'Ecuyer, David Munger, Boris Oreshkin, and Richard Simard
+#'     (2017). Random numbers for parallel computers: Requirements and methods,
+#'     with emphasis on GPUs. Mathematics and Computers in Simulation,
+#'     135, 3-17. \doi{10.1016/j.matcom.2016.05.005}.
 tar_seed_create <- function(name, global_seed = NULL) {
   if (is.null(global_seed)) {
     global_seed <- tar_options$get_seed()
@@ -67,13 +58,6 @@ tar_seed_create <- function(name, global_seed = NULL) {
   if (is.null(global_seed) || anyNA(global_seed)) {
     return(NA_integer_)
   }
-  name <- as.character(name)
-  hash <- digest::digest(
-    object = name,
-    algo = "sha512",
-    serialize = FALSE,
-    file = FALSE,
-    seed = 0L
-  )
-  digest::digest2int(x = hash, seed = global_seed)
+  x <- list(as.character(name), as.integer(global_seed))
+  secretbase::sha3(x = x, bits = 32L, convert = NA)
 }