+++ NEWS.md
@@ -1,4 +1,4 @@
-# targets 1.4.1.9003 (development)
+# targets 1.5.0
 
 ## Invalidating changes